// apps/webapp/src/app/core/api/models/tenant-membership-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RoleDto } from '../models/role-dto';
import { TenantDto } from '../models/tenant-dto';
import { UserDto } from '../models/user-dto';
export interface TenantMembershipDto {

  /**
   * Whether the membership is active.
   */
  is_active: boolean;

  /**
   * The unique identifier of the membership.
   */
  membership_id: string;

  /**
   * Custom permissions/overrides for this member.
   */
  permissions?: {
};

  /**
   * Role details
   */
  role?: RoleDto;

  /**
   * The role ID assigned to the user in this tenant.
   */
  role_id: string;

  /**
   * Tenant details
   */
  tenant?: TenantDto;

  /**
   * The tenant ID.
   */
  tenant_id: string;

  /**
   * User details
   */
  user?: UserDto;

  /**
   * The user ID.
   */
  user_id: string;
}