// apps/webapp/src/app/core/api/models/api-response-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ApiErrorPayloadDto } from '../models/api-error-payload-dto';
import { ApiMetaDto } from '../models/api-meta-dto';
export interface ApiResponseDto {
  data?: {
};
  error?: ApiErrorPayloadDto;
  message: string;
  meta: ApiMetaDto;
  statusCode: number;
  success: boolean;
}